+++ examples/c_api/tiledb_kv_read.c
@@ -53,7 +53,8 @@ int main() {
 
   // Open a key-value store
   tiledb_kv_t* kv;
-  tiledb_kv_open(ctx, &kv, "my_kv", NULL, 0);
+  tiledb_kv_alloc(ctx, "my_kv", &kv);
+  tiledb_kv_open(ctx, kv, NULL, 0);
 
   // Get key-value item
   int key = 100;